python/requirements.txt
gcloud
PyMySQL
flask-compress
loompy